code_refactor: move images into appropriate folders and rename some files to full names instead of abbreviations
